refactor(api): serve routing rules and complexity config under /api/routing with backwards-compatible aliases
